Move graphql specific utilities into `utilities/graphql`
