+++ src/__tests__/local-state/general.ts
@@ -8,7 +8,7 @@ import { Operation } from '../../link/core/types';
 import { ApolloClient } from '../..';
 import { ApolloCache } from '../../cache/core/cache';
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
-import { hasDirectives } from '../../utilities/directives';
+import { hasDirectives } from '../../utilities/graphql/directives';
 
 describe('General functionality', () => {
   it('should not impact normal non-@client use', () => {
