+++ src/cache/core/cache.ts
@@ -1,6 +1,6 @@
 import { DocumentNode } from 'graphql';
 
-import { getFragmentQueryDocument } from '../../utilities/fragments';
+import { getFragmentQueryDocument } from '../../utilities/graphql/fragments';
 import { DataProxy } from './types/DataProxy';
 import { Cache } from './types/Cache';
 import { justTypenameQuery, queryFromPojo, fragmentFromPojo } from './utils';
