+++ src/core/ObservableQuery.ts
@@ -4,7 +4,7 @@ import { invariant, InvariantError } from 'ts-invariant';
 import { isEqual } from '../utilities/util/isEqual';
 import { tryFunctionOrLogError } from '../utilities/util/errorHandling';
 import { cloneDeep } from '../utilities/util/cloneDeep';
-import { getOperationDefinition } from '../utilities/getFromAST';
+import { getOperationDefinition } from '../utilities/graphql/getFromAST';
 import { NetworkStatus, isNetworkRequestInFlight } from './networkStatus';
 import { Observable, Observer, Subscription } from '../util/Observable';
 import { ApolloError } from '../errors/ApolloError';
