+++ src/link/utils/transformOperation.ts
@@ -1,5 +1,5 @@
 import { GraphQLRequest, Operation } from '../core/types';
-import { getOperationName } from '../../utilities/getFromAST';
+import { getOperationName } from '../../utilities/graphql/getFromAST';
 
 export function transformOperation(operation: GraphQLRequest): GraphQLRequest {
   const transformedOperation: GraphQLRequest = {
