--- src/utilities/declarations.d.ts
@@ -1,1 +1,0 @@
-declare module 'fast-json-stable-stringify';
