+++ src/utilities/graphql/getFromAST.ts
@@ -7,7 +7,7 @@ import {
 
 import { invariant, InvariantError } from 'ts-invariant';
 
-import { assign } from './util/assign';
+import { assign } from '../util/assign';
 
 import { valueToObjectRepresentation, JsonValue } from './storeUtils';
